pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "biopython>=1.85",
    "email-validator>=2.2.0",
    "flasgger>=0.9.7.1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "prometheus-client>=0.21.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
    "sqlalchemy>=2.0.41",
    "werkzeug>=3.1.3",
]
